pipe/signal-prop: reshape mp3 package and update mod
