+++ go.mod
@@ -3,14 +3,17 @@ module github.com/pipelined/phono
 require (
 	github.com/go-audio/audio v0.0.0-20181013203223-7b2a6ca21480
 	github.com/go-audio/wav v0.0.0-20181013172942-de841e69b884
+	github.com/google/pprof v0.0.0-20190109223431-e84dfd68c163 // indirect
 	github.com/gordonklaus/portaudio v0.0.0-20180817120803-00e7307ccd93
 	github.com/hajimehoshi/go-mp3 v0.1.1
+	github.com/ianlancetaylor/demangle v0.0.0-20181102032728-5e5cf60278f6 // indirect
 	github.com/pipelined/vst2 v0.2.0
 	github.com/rs/xid v1.2.1
 	github.com/sirupsen/logrus v1.3.0
 	github.com/stretchr/testify v1.3.0
 	github.com/viert/lame v0.0.0-20190110105750-70ad2f7ac5f2
 	go.uber.org/goleak v0.10.0
+	golang.org/x/arch v0.0.0-20181203225421-5a4828bb7045 // indirect
 	golang.org/x/crypto v0.0.0-20190103213133-ff983b9c42bc // indirect
 	golang.org/x/sys v0.0.0-20190109145017-48ac38b7c8cb // indirect
 )
